A3/requirements.txt
einops==0.4.0
librosa==0.9.1
matplotlib==3.5.1
numpy==1.21.0
PySimpleGUI==4.60.1
pytorch_lightning==1.6.3
sounddevice==0.4.4
torchmetrics==0.8.2
validators==0.19.0
wandb==0.12.14
einops